requirements/local.txt
-r base.txt

psycopg2-binary==2.9.9
flake8==4.0.1
black==21.12b0
isort==5.9.3
watchfiles==0.18.1
argh==0.26.2
PyYAML==6.0
watchdog==2.1.6

pytest-django==4.5.2

pytest-factoryboy==2.5.1

Faker==18.3.1

pytest-cov==4.0.0